requirements/development.txt
-r testing.txt

docker-compose
pandas_datareader
plotly
